+++ asyncz/contrib/dashboard/config.py
@@ -36,9 +36,7 @@ class DashboardConfig:
     description: str = "A simple dashboard for monitoring Asyncz tasks."
     """A brief description of the dashboard's purpose."""
 
-    favicon: str = (
-        "https://raw.githubusercontent.com/dymmond/asyncz/refs/heads/main/docs/statics/favicon.ico"
-    )
+    favicon: str = "/static/favicon.ico"
     """URL path or external URL for the favicon."""
 
     dashboard_url_prefix: str = "/dashboard"